requirements.txt
astropy==5.0
astropy_healpix==0.6
celery==5.2.0
colorama==0.4.5
Django==4.0.4
django-extensions==3.1.5
django_filter==21.1
django_import_export==2.6.1
django-widget-tweaks==1.4.9
fastparquet==0.7.2
import_export==0.2.67.dev6
jupyterlab==3.4.5
MarkupPy==1.14
mod-wsgi==4.9.0
numpy==1.21.4
odfpy==1.4.1
openpyxl==3.0.9
pandas==1.4.1
pipdeptree==2.3.1
pipreqs==0.4.11
psycopg2-binary==2.9.2
pyarrow==6.0.1
pydotplus==2.0.2
pytest==7.2.0
python-decouple==3.5
sphinx-material==0.0.35
SQLAlchemy==1.4.27
stack-data==0.5.0
Unidecode==1.3.4
wheel==0.37.0
xlrd==2.0.1
xlwt==1.3.0
